Show proper title when asking for password on the backup restoration
Currently we show a popup with the same title and message both when we store/encrypt and restore/decrypt the backup.
This PR simply changes the title and clears the message in the latter case, so it won't confuse the user.
I still think it's not perfect - the popup asks for the password even if none is needed - but this is how it works
on iOS as well.